JP-2707: Fix failures in RunningComServerChangesTest

+++ comu/com.energyict.mdc.engine/src/test/java/com/energyict/mdc/engine/impl/core/mocks/MockOutboundComPort.java
@@ -24,7 +24,7 @@ public class MockOutboundComPort extends MockComPort implements OutboundComPort 
         super(comServer, name);
     }
 
-    public MockOutboundComPort (ComServer comServer, int id, String name) {
+    public MockOutboundComPort (ComServer comServer, long id, String name) {
         super(comServer, id, name);
     }
